requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
Authlib==1.3.1
babel==2.16.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
courlan==1.3.2
cryptography==44.0.0
dateparser==1.2.0
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.1.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpath-python==1.0.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.1
lxml==5.3.0
lxml_html_clean==0.4.1
matplotlib-inline==0.1.7
mechanize==0.4.10
mistralai==1.3.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
openai==1.59.8
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
selenium==4.27.1
sendgrid==6.11.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starkbank-ecdsa==2.2.0
tld==0.13
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
validators==0.34.0
wcwidth==0.2.13
weaviate==0.1.2
weaviate-client==4.10.4
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
